https://bugzilla.gnome.org/show_bug.cgi?id=760149

+++ shared/nm-macros-internal.h
@@ -208,7 +208,7 @@ nm_clear_g_source (guint *id)
 }
 
 static inline gboolean
-nm_clear_g_signal_handler (gpointer self, guint *id)
+nm_clear_g_signal_handler (gpointer self, gulong *id)
 {
 	if (id && *id) {
 		g_signal_handler_disconnect (self, *id);
